LayoutTests/http/tests/security/contentSecurityPolicy/1.1/frame-ancestors/frame-ancestors-nested-cross-in-cross-url-allow-expected.txt

CONSOLE MESSAGE: line 60: IFrame load event fired: the IFrame's location is 'http://localhost:8080/security/contentSecurityPolicy/resources/frame-ancestors.pl?policy=http://127.0.0.1:8000%20http://localhost:8080'.
A 'frame-ancestors' CSP directive with a URL value should block or allow rendering in nested frames as appropriate.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS The inner IFrame passed.
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Testing a cross-origin child with a policy of "http://127.0.0.1:8000 http://localhost:8080" nested in a cross-origin parent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS The IFrame should not have been blocked. It wasn't.


--------
Frame: '<!--framePath //<!--frame0-->/<!--frame0-->-->'
--------
This is an IFrame sending a Content Security Policy header containing "frame-ancestors http://127.0.0.1:8000 http://localhost:8080".
